testsets in output folder

+++ src/synthesis/synthesis_both_sides.m
@@ -27,9 +27,9 @@ function im_synth = synthesis_both_sides(disp_left,disp_right,leftrgb,rightrgb,p
     m=repmat(rowvls,size(leftrgb,1),1);
     m=int16(m);
     
-    new_x=m-disp_left.*p;
+    new_x=m-int16(disp_left).*p;
     new_x(new_x<1 | new_x > size(leftrgb,2))=m(new_x<1 | new_x > size(leftrgb,2));
-    new_x=uint16(new_x);
+    new_x=int16(new_x);
     
     for i=1:size(leftrgb,1)
         for j=1:size(leftrgb,2)
@@ -44,7 +44,7 @@ function im_synth = synthesis_both_sides(disp_left,disp_right,leftrgb,rightrgb,p
     m=repmat(rowvls,size(rightrgb,1),1);
     m=int16(m);
     
-    new_x=m+disp_right.*(1-p);
+    new_x=m+int16(disp_right).*(1-p);
     new_x(new_x<1 | new_x > size(rightrgb,2))=m(new_x<1 | new_x > size(rightrgb,2));
     new_x=int16(new_x);
